CVE-2020-10379 hard copy src/libImaging/TiffDecode.c from tag 7.1.0
I just hard copied TiffDecode from the fixed version, this likely means we have other fixes in there as well, more than just this CVE, but that's basically good thing. The reason I had to hard copy it is that the cherry-pick wasn't really working because the code had changed so much.